fix: rename `PortalMenu` file to `ContextMenu`
