+++ packages/shared/src/components/tooltips/ContextMenuItem.tsx
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 import Link from 'next/link';
 import React, { ReactElement } from 'react';
 import ConditionalWrapper from '../ConditionalWrapper';
-import { MenuItemProps } from '../fields/PortalMenu';
+import { MenuItemProps } from '../fields/ContextMenu';
 import { IconSize } from '../Icon';
 
 export interface ContextMenuItemProps
